+++ controllers/keda/scaledobject_controller_test.go
@@ -1179,7 +1179,7 @@ var _ = Describe("ScaledObjectController", func() {
 				return metav1.ConditionTrue
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
-		}, 5*time.Second).Should(Or(Equal(metav1.ConditionFalse), Equal(metav1.ConditionUnknown)))
+		}, 5*time.Second).Should(Equal(metav1.ConditionFalse))
 
 		// set annotation
 		Eventually(func() error {
@@ -1204,7 +1204,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(2 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionTrue))
@@ -1268,7 +1268,7 @@ var _ = Describe("ScaledObjectController", func() {
 				return metav1.ConditionTrue
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
-		}, 5*time.Second).Should(Or(Equal(metav1.ConditionFalse), Equal(metav1.ConditionUnknown)))
+		}, 5*time.Second).Should(Equal(metav1.ConditionFalse))
 
 		// set annotation to true at first
 		Eventually(func() error {
@@ -1293,7 +1293,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(2 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionTrue))
@@ -1322,7 +1322,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionTrue
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(2 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionFalse))
@@ -1401,7 +1401,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).Should(Equal(metav1.ConditionTrue))
@@ -1410,7 +1410,7 @@ var _ = Describe("ScaledObjectController", func() {
 		Eventually(func() metav1.ConditionStatus {
 			err := k8sClient.Get(context.Background(), types.NamespacedName{Name: soName, Namespace: "default"}, so)
 			if err != nil {
-				return metav1.ConditionUnknown
+				return metav1.ConditionFalse
 			}
 			return so.Status.Conditions.GetPausedCondition().Status
 		}).WithTimeout(1 * time.Minute).WithPolling(10 * time.Second).Should(Equal(metav1.ConditionTrue))